Title: 선분 위의 점, Time: -1ms, Memory: -1MB, Status: Compile Error - CodeTree

+++ 240306/선분 위의 점/point-on-the-line-segment.java
@@ -38,7 +38,7 @@ public class Main {
         int minIdx = n+1;
 
         while(left <= right){
-            int  mid = (left + right) / 2;
+            Long  mid = (left + right) / 2;
 
             if(arr[mid] >= target){
                 right = mid - 1;
